solved dashed/dotted problem

+++ src/python/ccpn/ui/gui/widgets/SpectrumToolBar.py
@@ -32,7 +32,6 @@ from ccpn.ui.gui.widgets.ToolBar import ToolBar
 from functools import partial
 from ccpn.core.lib.Notifiers import Notifier
 from collections import OrderedDict
-from ccpn.ui.gui.widgets.MessageDialog import showWarning
 
 
 class SpectrumToolBar(ToolBar):
@@ -64,11 +63,11 @@ class SpectrumToolBar(ToolBar):
       spectrumView = self.widget.project.getByPid(action.spectrumViewPid)
       newSpectrumViewsOrder.append(spectrumView)
 
-    # self.widget.project.blankNotification()
+    self.widget.project.blankNotification()
     # newIndex = [newIndex.index(ii) for ii in self.widget.getOrderedSpectrumViewsIndex()]
     newIndex = [self.widget.getOrderedSpectrumViewsIndex()[ii] for ii in newIndex]
     self.widget.setOrderedSpectrumViewsIndex(newIndex)
-    # self.widget.project.unblankNotification()
+    self.widget.project.unblankNotification()
 
     # defaults = OrderedDict((('newIndex', None),))
     #
@@ -212,7 +211,7 @@ class SpectrumToolBar(ToolBar):
     """
     Removes the spectrum from the display and its button from the toolbar.
     """
-    # TODO:ED this needs patching in to the spectrumView.delete()
+
     # remove the item from the toolbar
     self.removeAction(button.actions()[0])
 
@@ -221,16 +220,32 @@ class SpectrumToolBar(ToolBar):
     stripUpdateList = []
     for spectrumView in self.widget.spectrumViews:
       if spectrumView._apiDataSource == key:
-        index = self.widget.spectrumViews.index(spectrumView)
-        break
-    else:
-      showWarning('Spectrum not found in toolbar')
-      return
+        stripUpdateList.append(spectrumView)
+        # self.widget.removeOrderedSpectrumView(spectrumView)
 
-    # found value is spectrumView, index
+        # this spawns the creation of all orderedSpectra
+        # should be done on loading though
+        # spectrumView.strip.orderedSpectra()
 
-    spectrumView.delete()
-    # self.widget.removeOrderedSpectrumView(index)
+    for spectrumView in self.widget.spectrumViews:
+      if spectrumView._apiDataSource == key:
+
+        # delete the spectrumView
+        try:
+          # spectrumView._wrappedData.spectrumView.delete()
+          # if not spectrumView.isDeleted:
+          # self.widget.project.blankNotifications()
+          # self.widget.removeOrderedSpectrumView(spectrumView)
+          # self.widget.project.unblankNotifications()
+
+          spectrumView.delete()
+        except Exception as es:
+          pass
+
+    # for st in stripUpdateList:
+    # #   st.removeSpectrumView(None)
+    #   if not st.isDeleted:
+    #     self.widget.removeOrderedSpectrumView(st)
 
     # spawn a redraw of the GL windows
     from ccpn.ui.gui.lib.OpenGL.CcpnOpenGL import GLNotifier
@@ -321,8 +336,7 @@ class SpectrumToolBar(ToolBar):
       actionList = self.actions()
       try:
         # try and find the spectrumView in the orderedlist - for undo function
-        oldList = spectrumDisplay.spectrumViews
-        oldList = spectrumDisplay.orderedSpectrumViews(oldList)
+        oldList = spectrumDisplay.orderedSpectrumViews()
         oldIndex = oldList.index(spectrumView)
 
         if actionList and oldIndex < len(actionList):
@@ -368,7 +382,6 @@ class SpectrumToolBar(ToolBar):
     self.clear()
     for specView in spectrumViews:
 
-      # self._addSpectrumViewToolButtons(specView)
       spectrum = specView.spectrum
       spectrumName = spectrum.name
       if len(spectrumName) > 12:
